+++ CHANGELOG.md
@@ -17,17 +17,21 @@ This is our initial release of the plugin
 
 ### Added
 - Touch events 
-- Event on frame change "wpe/ScrollSequence/frame"
+- Event on frame change "wpg/ScrollSequence/frame"
 
-#### "wpe/ScrollSequence/frame"
+#### "wpg/ScrollSequence/frame"
 Every time the frame changes this event is dispatched on window.
 The Container ID and teh current frame are in ethe event detail
 
 eg.
 <pre>
-window.addEventListener('wpe/ScrollSequence/frame', e => { 
+window.addEventListener('wpg/ScrollSequence/frame', e => { 
     const {id, frame} = e.detail;
     console.log('id': id);
     console.log('frame': frame);
 })
 </pre>
+
+### Fixes
+- Renamed all incorrect namings of 'wpe' to 'wpg'
+- Renamed incorrect naming of 'toolbar' to 'tipbar'